@@ -1,15 +1,17 @@
-import { BrowserRouter as Router } from 'react-router-dom';
+import { BrowserRouter as Router } from "react-router-dom";
+import { FoodProvider } from "./contexts/FoodContext";
+import Routes from "./routes";
 
-import Routes from './routes';
-
-import GlobalStyle from './styles/global';
+import GlobalStyle from "./styles/global";
 
 const App = () => (
   <>
-    <GlobalStyle />
-    <Router>
-      <Routes />
-    </Router>
+    <FoodProvider>
+      <GlobalStyle />
+      <Router>
+        <Routes />
+      </Router>
+    </FoodProvider>
   </>
 );
 
